drm: bridge: samsung-dsim: Support multi-lane calculations
If there is more than one lane, the HFP, HBP, and HSA is calculated in bytes/pixel,
then they are divided amongst the different lanes with some additional overhead.
This is necesary to achieve higher resolutions while keeping the pixel clocks lower.
To match the NXP downstream kernel, round the HFP and HBP up and round the HSA down.

Signed-off-by: Adam Ford <aford173@gmail.com>