drm: bridge: samsung-dsim: Dynamically configure DPHY timing
NXP uses a lookup table to determine the various values for
the PHY Timing based on the clock rate.  Since the input
clock can be variable, the phy settings need to be variable
too.  Add an additional variable to the driver data to enable
this feature to prevent breaking boards that don't support it.

Signed-off-by: Adam Ford <aford173@gmail.com>