drm/bridge: Fix bridge disabling order
As explained by the comment of the fixed code, we need to find the next
bridge that hasn't set the pre_enable_prev_first flag to true. The code,
on the contrary, was doing the opposite, i. e., finding the next bridge
with this flag set to false.
So, the order of disabling the bridges couldn't be altered as required
by setting the pre_enable_prev_first flag to true.

Fixes: 4fb912e5e190 ("drm/bridge: Introduce pre_enable_prev_first to alter bridge init order")
Signed-off-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>