drm/bridge: Fix a use case in bridge disabling order
The patch fixes the code for finding the next bridge with the flag
pre_enable_prev_first set to false. In the case that this condition is
not met, meaning there is no subsequent bridge with the flag set to false,
the whole bridge list is traversed, invalidating the next variable.

The use of a new iteration variable (i. e. iter) ensures that the value
of the next variable is not invalidated in the described use case, a
variable that must then be used by the subsequent logic.

Fixes: 4fb912e5e190 ("drm/bridge: Introduce pre_enable_prev_first to alter bridge init order")
Co-developed-by: Michael Trimarchi <michael@amarulasolutions.com>
Signed-off-by: Michael Trimarchi <michael@amarulasolutions.com>
Signed-off-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>